Calculate distance as float instead of int (get some sub-cm accuracy)
Change serial output to use snprintf

+++ src/music.h
@@ -1,6 +1,6 @@
 #ifndef MUSIC_H
 #define MUSIC_H
 
-float frequency_from_distance(int distance, int maxDistance, double minFrequency, int rangeSize);
+float frequency_from_distance(float distance, float maxDistance, double minFrequency, int rangeSize);
 
 #endif
